requirements.txt
requests
pyewts
uuid
PyYaml
pathlib
rdflib